move analytics callback to Renderer
